+++ CHANGELOG.md
@@ -5,7 +5,7 @@
     is pressed.
   * [Added] ext-multi is now deployed with jquery.fancytree-all.js (still experimental)
   * [Added] tree.activateKey(key, opts) now has an `opts` argument
-  * [Added] `noData` option (bool, string, or callback)
+  * [Added] `nodata` option (bool, string, or callback)
   * [Added] ext-table `mergeStatusColumns` option
   * [Fixed] ES6 import dependency on jquery for jquery.fancytree.ui-deps.js
   * [Fixed] #863 setActive() sometimes does not scroll node into view